+++ x-pack/plugin/core/src/test/java/org/elasticsearch/xpack/core/ilm/LifecycleExecutionStateTests.java
@@ -71,6 +71,69 @@ public class LifecycleExecutionStateTests extends ESTestCase {
             LifecycleExecutionStateTests::mutate);
     }
 
+    public void testGetCurrentStepKey() {
+        LifecycleExecutionState.Builder lifecycleState = LifecycleExecutionState.builder();
+        Step.StepKey stepKey = LifecycleExecutionState.getCurrentStepKey(lifecycleState.build());
+        assertNull(stepKey);
+
+        String phase = randomAlphaOfLength(20);
+        String action = randomAlphaOfLength(20);
+        String step = randomAlphaOfLength(20);
+        LifecycleExecutionState.Builder lifecycleState2 = LifecycleExecutionState.builder();
+        lifecycleState2.setPhase(phase);
+        lifecycleState2.setAction(action);
+        lifecycleState2.setStep(step);
+        stepKey = LifecycleExecutionState.getCurrentStepKey(lifecycleState2.build());
+        assertNotNull(stepKey);
+        assertEquals(phase, stepKey.getPhase());
+        assertEquals(action, stepKey.getAction());
+        assertEquals(step, stepKey.getName());
+
+        phase = randomAlphaOfLength(20);
+        action = randomAlphaOfLength(20);
+        step = null;
+        LifecycleExecutionState.Builder lifecycleState3 = LifecycleExecutionState.builder();
+        lifecycleState3.setPhase(phase);
+        lifecycleState3.setAction(action);
+        lifecycleState3.setStep(step);
+        AssertionError error3 = expectThrows(AssertionError.class,
+            () -> LifecycleExecutionState.getCurrentStepKey(lifecycleState3.build()));
+        assertEquals("Current phase is not empty: " + phase, error3.getMessage());
+
+        phase = null;
+        action = randomAlphaOfLength(20);
+        step = null;
+        LifecycleExecutionState.Builder lifecycleState4 = LifecycleExecutionState.builder();
+        lifecycleState4.setPhase(phase);
+        lifecycleState4.setAction(action);
+        lifecycleState4.setStep(step);
+        AssertionError error4 = expectThrows(AssertionError.class,
+            () -> LifecycleExecutionState.getCurrentStepKey(lifecycleState4.build()));
+        assertEquals("Current action is not empty: " + action, error4.getMessage());
+
+        phase = null;
+        action = randomAlphaOfLength(20);
+        step = randomAlphaOfLength(20);
+        LifecycleExecutionState.Builder lifecycleState5 = LifecycleExecutionState.builder();
+        lifecycleState5.setPhase(phase);
+        lifecycleState5.setAction(action);
+        lifecycleState5.setStep(step);
+        AssertionError error5 = expectThrows(AssertionError.class,
+            () -> LifecycleExecutionState.getCurrentStepKey(lifecycleState5.build()));
+        assertNull(error5.getMessage());
+
+        phase = null;
+        action = null;
+        step = randomAlphaOfLength(20);
+        LifecycleExecutionState.Builder lifecycleState6 = LifecycleExecutionState.builder();
+        lifecycleState6.setPhase(phase);
+        lifecycleState6.setAction(action);
+        lifecycleState6.setStep(step);
+        AssertionError error6 = expectThrows(AssertionError.class,
+            () -> LifecycleExecutionState.getCurrentStepKey(lifecycleState6.build()));
+        assertNull(error6.getMessage());
+    }
+
     private static LifecycleExecutionState mutate(LifecycleExecutionState toMutate) {
         LifecycleExecutionState.Builder newState = LifecycleExecutionState.builder(toMutate);
         boolean changed = false;
