+++ x-pack/plugin/ilm/qa/rest/src/test/resources/rest-api-spec/test/ilm/40_explain_lifecycle.yml
@@ -115,9 +115,9 @@ teardown:
   - match: { indices.my_index.step: "complete" }
   - is_true: indices.my_index.phase_time_millis
   - is_true: indices.my_index.age
+  - is_true: indices.my_index.phase_execution
   - is_false: indices.my_index.failed_step
   - is_false: indices.my_index.step_info
-  - is_false: indices.my_index.phase_execution
 
   - is_false: indices.my_index2
   - is_false: indices.another_index
@@ -139,9 +139,9 @@ teardown:
   - match: { indices.my_index.step: "complete" }
   - is_true: indices.my_index.phase_time_millis
   - is_true: indices.my_index.age
+  - is_true: indices.my_index.phase_execution
   - is_false: indices.my_index.failed_step
   - is_false: indices.my_index.step_info
-  - is_false: indices.my_index.phase_execution
 
   - is_true: indices.my_index2.managed
   - match: { indices.my_index2.index: "my_index2" }
@@ -151,9 +151,9 @@ teardown:
   - match: { indices.my_index2.step: "complete" }
   - is_true: indices.my_index2.phase_time_millis
   - is_true: indices.my_index2.age
+  - is_true: indices.my_index2.phase_execution
   - is_false: indices.my_index2.failed_step
   - is_false: indices.my_index2.step_info
-  - is_false: indices.my_index2.phase_execution
 
   - is_false: indices.another_index
   - is_false: indices.unmanaged_index
@@ -178,9 +178,9 @@ teardown:
   - match: { indices.my_index.step: "complete" }
   - is_true: indices.my_index.phase_time_millis
   - is_true: indices.my_index.age
+  - is_true: indices.my_index.phase_execution
   - is_false: indices.my_index.failed_step
   - is_false: indices.my_index.step_info
-  - is_false: indices.my_index.phase_execution
 
   - is_true: indices.my_index2.managed
   - match: { indices.my_index2.index: "my_index2" }
@@ -190,9 +190,9 @@ teardown:
   - match: { indices.my_index2.step: "complete" }
   - is_true: indices.my_index2.phase_time_millis
   - is_true: indices.my_index2.age
+  - is_true: indices.my_index2.phase_execution
   - is_false: indices.my_index2.failed_step
   - is_false: indices.my_index2.step_info
-  - is_false: indices.my_index2.phase_execution
 
   - is_true: indices.another_index.managed
   - match: { indices.another_index.index: "another_index" }
@@ -202,9 +202,9 @@ teardown:
   - match: { indices.another_index.step: "complete" }
   - is_true: indices.another_index.phase_time_millis
   - is_true: indices.another_index.age
+  - is_true: indices.another_index.phase_execution
   - is_false: indices.another_index.failed_step
   - is_false: indices.another_index.step_info
-  - is_false: indices.another_index.phase_execution
 
   - match: { indices.unmanaged_index.index: "unmanaged_index" }
   - is_false: indices.unmanaged_index.managed
