+++ x-pack/plugin/ilm/src/main/java/org/elasticsearch/xpack/ilm/ExecuteStepsUpdateTask.java
@@ -16,7 +16,6 @@ import org.elasticsearch.common.xcontent.ToXContentObject;
 import org.elasticsearch.index.Index;
 import org.elasticsearch.xpack.core.ilm.ClusterStateActionStep;
 import org.elasticsearch.xpack.core.ilm.ClusterStateWaitStep;
-import org.elasticsearch.xpack.core.ilm.LifecycleExecutionState;
 import org.elasticsearch.xpack.core.ilm.Step;
 import org.elasticsearch.xpack.core.ilm.TerminalPolicyStep;
 
@@ -78,8 +77,7 @@ public class ExecuteStepsUpdateTask extends ClusterStateUpdateTask {
             // This index doesn't exist any more, there's nothing to execute currently
             return currentState;
         }
-        Step registeredCurrentStep = IndexLifecycleRunner.getCurrentStep(policyStepsRegistry, policy, indexMetaData,
-            LifecycleExecutionState.fromIndexMetadata(indexMetaData));
+        Step registeredCurrentStep = IndexLifecycleRunner.getCurrentStep(policyStepsRegistry, policy, indexMetaData);
         if (currentStep.equals(registeredCurrentStep)) {
             ClusterState state = currentState;
             // We can do cluster state steps all together until we
@@ -103,8 +101,8 @@ public class ExecuteStepsUpdateTask extends ClusterStateUpdateTask {
                         return state;
                     } else {
                         logger.trace("[{}] moving cluster state to next step [{}]", index.getName(), nextStepKey);
-                        state = IndexLifecycleRunner.moveClusterStateToNextStep(index, state, currentStep.getKey(),
-                            nextStepKey, nowSupplier, false);
+                        state = IndexLifecycleTransition.moveClusterStateToStep(index, state, nextStepKey, nowSupplier,
+                            policyStepsRegistry, false);
                     }
                 } else {
                     // set here to make sure that the clusterProcessed knows to execute the
@@ -130,8 +128,8 @@ public class ExecuteStepsUpdateTask extends ClusterStateUpdateTask {
                         if (currentStep.getNextStepKey() == null) {
                             return state;
                         } else {
-                            state = IndexLifecycleRunner.moveClusterStateToNextStep(index, state, currentStep.getKey(),
-                                currentStep.getNextStepKey(), nowSupplier, false);
+                            state = IndexLifecycleTransition.moveClusterStateToStep(index, state,
+                                currentStep.getNextStepKey(), nowSupplier, policyStepsRegistry,false);
                         }
                     } else {
                         logger.trace("[{}] condition not met ({}) [{}], returning existing state",
@@ -145,7 +143,7 @@ public class ExecuteStepsUpdateTask extends ClusterStateUpdateTask {
                         if (stepInfo == null) {
                             return state;
                         } else {
-                            return IndexLifecycleRunner.addStepInfoToClusterState(index, state, stepInfo);
+                            return IndexLifecycleTransition.addStepInfoToClusterState(index, state, stepInfo);
                         }
                     }
                 }
