+++ x-pack/plugin/ilm/src/main/java/org/elasticsearch/xpack/ilm/MoveToNextStepUpdateTask.java
@@ -28,15 +28,18 @@ public class MoveToNextStepUpdateTask extends ClusterStateUpdateTask {
     private final Step.StepKey currentStepKey;
     private final Step.StepKey nextStepKey;
     private final LongSupplier nowSupplier;
+    private final PolicyStepsRegistry stepRegistry;
     private final Consumer<ClusterState> stateChangeConsumer;
 
     public MoveToNextStepUpdateTask(Index index, String policy, Step.StepKey currentStepKey, Step.StepKey nextStepKey,
-                                    LongSupplier nowSupplier, Consumer<ClusterState> stateChangeConsumer) {
+                                    LongSupplier nowSupplier, PolicyStepsRegistry stepRegistry,
+                                    Consumer<ClusterState> stateChangeConsumer) {
         this.index = index;
         this.policy = policy;
         this.currentStepKey = currentStepKey;
         this.nextStepKey = nextStepKey;
         this.nowSupplier = nowSupplier;
+        this.stepRegistry = stepRegistry;
         this.stateChangeConsumer = stateChangeConsumer;
     }
 
@@ -66,9 +69,9 @@ public class MoveToNextStepUpdateTask extends ClusterStateUpdateTask {
         Settings indexSettings = indexMetaData.getSettings();
         LifecycleExecutionState indexILMData = LifecycleExecutionState.fromIndexMetadata(currentState.getMetaData().index(index));
         if (policy.equals(LifecycleSettings.LIFECYCLE_NAME_SETTING.get(indexSettings))
-            && currentStepKey.equals(IndexLifecycleRunner.getCurrentStepKey(indexILMData))) {
+            && currentStepKey.equals(LifecycleExecutionState.getCurrentStepKey(indexILMData))) {
             logger.trace("moving [{}] to next step ({})", index.getName(), nextStepKey);
-            return IndexLifecycleRunner.moveClusterStateToNextStep(index, currentState, currentStepKey, nextStepKey, nowSupplier, false);
+            return IndexLifecycleTransition.moveClusterStateToStep(index, currentState, nextStepKey, nowSupplier, stepRegistry, false);
         } else {
             // either the policy has changed or the step is now
             // not the same as when we submitted the update task. In
