+++ x-pack/plugin/ilm/src/test/java/org/elasticsearch/xpack/ilm/IndexLifecycleRunnerTests.java
@@ -5,7 +5,8 @@
  */
 package org.elasticsearch.xpack.ilm;
 
-import org.elasticsearch.ElasticsearchException;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.elasticsearch.Version;
 import org.elasticsearch.client.Client;
 import org.elasticsearch.cluster.ClusterName;
@@ -19,7 +20,6 @@ import org.elasticsearch.cluster.service.ClusterService;
 import org.elasticsearch.common.Strings;
 import org.elasticsearch.common.bytes.BytesReference;
 import org.elasticsearch.common.settings.Settings;
-import org.elasticsearch.common.settings.Settings.Builder;
 import org.elasticsearch.common.unit.TimeValue;
 import org.elasticsearch.common.xcontent.NamedXContentRegistry;
 import org.elasticsearch.common.xcontent.ToXContent;
@@ -32,7 +32,6 @@ import org.elasticsearch.test.ESTestCase;
 import org.elasticsearch.test.client.NoOpClient;
 import org.elasticsearch.threadpool.TestThreadPool;
 import org.elasticsearch.threadpool.ThreadPool;
-import org.elasticsearch.xpack.core.ilm.AbstractStepTestCase;
 import org.elasticsearch.xpack.core.ilm.AsyncActionStep;
 import org.elasticsearch.xpack.core.ilm.AsyncWaitStep;
 import org.elasticsearch.xpack.core.ilm.ClusterStateActionStep;
@@ -63,6 +62,7 @@ import org.mockito.Mockito;
 
 import java.io.IOException;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
@@ -75,8 +75,6 @@ import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicLong;
 import java.util.function.BiFunction;
-import java.util.function.Function;
-import java.util.stream.Collectors;
 
 import static java.util.stream.Collectors.toList;
 import static org.elasticsearch.test.hamcrest.ElasticsearchAssertions.awaitLatch;
@@ -85,7 +83,6 @@ import static org.elasticsearch.xpack.core.ilm.LifecyclePolicyTestsUtils.newTest
 import static org.hamcrest.Matchers.containsString;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.is;
-import static org.hamcrest.Matchers.nullValue;
 import static org.mockito.Matchers.any;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.times;
@@ -110,55 +107,6 @@ public class IndexLifecycleRunnerTests extends ESTestCase {
     @After
     public void shutdown() {
         threadPool.shutdownNow();
-    }
-
-    /** A real policy steps registry where getStep can be overridden so that JSON doesn't have to be parsed */
-    private class MockPolicyStepsRegistry extends PolicyStepsRegistry {
-        private BiFunction<IndexMetaData, StepKey, Step> fn = null;
-
-        MockPolicyStepsRegistry(SortedMap<String, LifecyclePolicyMetadata> lifecyclePolicyMap, Map<String, Step> firstStepMap,
-                                Map<String, Map<StepKey, Step>> stepMap, NamedXContentRegistry xContentRegistry, Client client) {
-            super(lifecyclePolicyMap, firstStepMap, stepMap, xContentRegistry, client);
-        }
-
-        public void setResolver(BiFunction<IndexMetaData, StepKey, Step> fn) {
-            this.fn = fn;
-        }
-
-        @Override
-        public Step getStep(IndexMetaData indexMetaData, StepKey stepKey) {
-            if (fn == null) {
-                logger.info("--> retrieving step {}", stepKey);
-                return super.getStep(indexMetaData, stepKey);
-            } else {
-                logger.info("--> returning mock step");
-                return fn.apply(indexMetaData, stepKey);
-            }
-        }
-    }
-
-    private MockPolicyStepsRegistry createOneStepPolicyStepRegistry(String policyName, Step step) {
-        return createOneStepPolicyStepRegistry(policyName, step, "test");
-    }
-
-    private MockPolicyStepsRegistry createOneStepPolicyStepRegistry(String policyName, Step step, String indexName) {
-        LifecyclePolicy policy = new LifecyclePolicy(policyName, new HashMap<>());
-        SortedMap<String, LifecyclePolicyMetadata> lifecyclePolicyMap = new TreeMap<>();
-        lifecyclePolicyMap.put(policyName, new LifecyclePolicyMetadata(policy, new HashMap<>(), 1, 1));
-        Map<String, Step> firstStepMap = new HashMap<>();
-        firstStepMap.put(policyName, step);
-        Map<String, Map<StepKey, Step>> stepMap = new HashMap<>();
-        Map<StepKey, Step> policySteps = new HashMap<>();
-        policySteps.put(step.getKey(), step);
-        stepMap.put(policyName, policySteps);
-        Map<Index, List<Step>> indexSteps = new HashMap<>();
-        List<Step> steps = new ArrayList<>();
-        steps.add(step);
-        Index index = new Index(indexName, indexName + "uuid");
-        indexSteps.put(index, steps);
-        Client client = mock(Client.class);
-        when(client.settings()).thenReturn(Settings.EMPTY);
-        return new MockPolicyStepsRegistry(lifecyclePolicyMap, firstStepMap, stepMap, REGISTRY, client);
     }
 
     public void testRunPolicyTerminalPolicyStep() {
@@ -295,7 +243,7 @@ public class IndexLifecycleRunnerTests extends ESTestCase {
         StepKey nextStepKey = new StepKey("phase", "action", "next_cluster_state_action_step");
         MockClusterStateActionStep step = new MockClusterStateActionStep(stepKey, nextStepKey);
         MockClusterStateActionStep nextStep = new MockClusterStateActionStep(nextStepKey, null);
-        MockPolicyStepsRegistry stepRegistry = createOneStepPolicyStepRegistry(policyName, step);
+        MockPolicyStepsRegistry stepRegistry = createMultiStepPolicyStepRegistry(policyName, Arrays.asList(step, nextStep));
         stepRegistry.setResolver((i, k) -> {
             if (stepKey.equals(k)) {
                 return step;
@@ -481,7 +429,7 @@ public class IndexLifecycleRunnerTests extends ESTestCase {
         StepKey nextStepKey = new StepKey("phase", "action", "async_action_step");
         MockClusterStateActionStep step = new MockClusterStateActionStep(stepKey, nextStepKey);
         MockAsyncActionStep nextStep = new MockAsyncActionStep(nextStepKey, null);
-        MockPolicyStepsRegistry stepRegistry = createOneStepPolicyStepRegistry(policyName, step);
+        MockPolicyStepsRegistry stepRegistry = createMultiStepPolicyStepRegistry(policyName, Arrays.asList(step, nextStep));
         stepRegistry.setResolver((i, k) -> {
             if (stepKey.equals(k)) {
                 return step;
@@ -696,65 +644,6 @@ public class IndexLifecycleRunnerTests extends ESTestCase {
         Mockito.verifyNoMoreInteractions(clusterService);
     }
 
-    public void testGetCurrentStepKey() {
-        LifecycleExecutionState.Builder lifecycleState = LifecycleExecutionState.builder();
-        StepKey stepKey = IndexLifecycleRunner.getCurrentStepKey(lifecycleState.build());
-        assertNull(stepKey);
-
-        String phase = randomAlphaOfLength(20);
-        String action = randomAlphaOfLength(20);
-        String step = randomAlphaOfLength(20);
-        LifecycleExecutionState.Builder lifecycleState2 = LifecycleExecutionState.builder();
-        lifecycleState2.setPhase(phase);
-        lifecycleState2.setAction(action);
-        lifecycleState2.setStep(step);
-        stepKey = IndexLifecycleRunner.getCurrentStepKey(lifecycleState2.build());
-        assertNotNull(stepKey);
-        assertEquals(phase, stepKey.getPhase());
-        assertEquals(action, stepKey.getAction());
-        assertEquals(step, stepKey.getName());
-
-        phase = randomAlphaOfLength(20);
-        action = randomAlphaOfLength(20);
-        step = null;
-        LifecycleExecutionState.Builder lifecycleState3 = LifecycleExecutionState.builder();
-        lifecycleState3.setPhase(phase);
-        lifecycleState3.setAction(action);
-        lifecycleState3.setStep(step);
-        AssertionError error3 = expectThrows(AssertionError.class, () -> IndexLifecycleRunner.getCurrentStepKey(lifecycleState3.build()));
-        assertEquals("Current phase is not empty: " + phase, error3.getMessage());
-
-        phase = null;
-        action = randomAlphaOfLength(20);
-        step = null;
-        LifecycleExecutionState.Builder lifecycleState4 = LifecycleExecutionState.builder();
-        lifecycleState4.setPhase(phase);
-        lifecycleState4.setAction(action);
-        lifecycleState4.setStep(step);
-        AssertionError error4 = expectThrows(AssertionError.class, () -> IndexLifecycleRunner.getCurrentStepKey(lifecycleState4.build()));
-        assertEquals("Current action is not empty: " + action, error4.getMessage());
-
-        phase = null;
-        action = randomAlphaOfLength(20);
-        step = randomAlphaOfLength(20);
-        LifecycleExecutionState.Builder lifecycleState5 = LifecycleExecutionState.builder();
-        lifecycleState5.setPhase(phase);
-        lifecycleState5.setAction(action);
-        lifecycleState5.setStep(step);
-        AssertionError error5 = expectThrows(AssertionError.class, () -> IndexLifecycleRunner.getCurrentStepKey(lifecycleState5.build()));
-        assertEquals(null, error5.getMessage());
-
-        phase = null;
-        action = null;
-        step = randomAlphaOfLength(20);
-        LifecycleExecutionState.Builder lifecycleState6 = LifecycleExecutionState.builder();
-        lifecycleState6.setPhase(phase);
-        lifecycleState6.setAction(action);
-        lifecycleState6.setStep(step);
-        AssertionError error6 = expectThrows(AssertionError.class, () -> IndexLifecycleRunner.getCurrentStepKey(lifecycleState6.build()));
-        assertEquals(null, error6.getMessage());
-    }
-
     public void testGetCurrentStep() {
         String policyName = "policy";
         StepKey firstStepKey = new StepKey("phase_1", "action_1", "step_1");
@@ -796,559 +685,16 @@ public class IndexLifecycleRunnerTests extends ESTestCase {
         PolicyStepsRegistry registry = new PolicyStepsRegistry(metas, firstStepMap, stepMap, REGISTRY, client);
 
         // First step is retrieved because there are no settings for the index
-        Step stepFromNoSettings = IndexLifecycleRunner.getCurrentStep(registry, policy.getName(), indexMetaData,
-            LifecycleExecutionState.builder().build());
+        IndexMetaData indexMetaDataWithNoKey = IndexMetaData.builder(index.getName())
+            .settings(indexSettings)
+            .putCustom(ILM_CUSTOM_METADATA_KEY, LifecycleExecutionState.builder().build().asMap())
+            .build();
+        Step stepFromNoSettings = IndexLifecycleRunner.getCurrentStep(registry, policy.getName(), indexMetaDataWithNoKey);
         assertEquals(firstStep, stepFromNoSettings);
 
         // The step that was written into the metadata is retrieved
-        Step currentStep = IndexLifecycleRunner.getCurrentStep(registry, policy.getName(), indexMetaData, lifecycleState.build());
+        Step currentStep = IndexLifecycleRunner.getCurrentStep(registry, policy.getName(), indexMetaData);
         assertEquals(step.getKey(), currentStep.getKey());
-    }
-
-    public void testMoveClusterStateToNextStep() {
-        String indexName = "my_index";
-        LifecyclePolicy policy = randomValueOtherThanMany(p -> p.getPhases().size() == 0,
-            () -> LifecyclePolicyTests.randomTestLifecyclePolicy("policy"));
-        Phase nextPhase = policy.getPhases().values().stream().findFirst().get();
-        List<LifecyclePolicyMetadata> policyMetadatas = Collections.singletonList(
-            new LifecyclePolicyMetadata(policy, Collections.emptyMap(), randomNonNegativeLong(), randomNonNegativeLong()));
-        StepKey currentStep = new StepKey("current_phase", "current_action", "current_step");
-        StepKey nextStep = new StepKey(nextPhase.getName(), "next_action", "next_step");
-        long now = randomNonNegativeLong();
-
-        // test going from null lifecycle settings to next step
-        ClusterState clusterState = buildClusterState(indexName,
-            Settings.builder()
-                .put(LifecycleSettings.LIFECYCLE_NAME, policy.getName()), LifecycleExecutionState.builder().build(), policyMetadatas);
-        Index index = clusterState.metaData().index(indexName).getIndex();
-        ClusterState newClusterState = IndexLifecycleRunner.moveClusterStateToNextStep(index, clusterState, currentStep, nextStep,
-                () -> now, false);
-        assertClusterStateOnNextStep(clusterState, index, currentStep, nextStep, newClusterState, now);
-
-        LifecycleExecutionState.Builder lifecycleState = LifecycleExecutionState.builder();
-        lifecycleState.setPhase(currentStep.getPhase());
-        lifecycleState.setAction(currentStep.getAction());
-        lifecycleState.setStep(currentStep.getName());
-        // test going from set currentStep settings to nextStep
-        Builder indexSettingsBuilder = Settings.builder()
-            .put(LifecycleSettings.LIFECYCLE_NAME, policy.getName());
-        if (randomBoolean()) {
-            lifecycleState.setStepInfo(randomAlphaOfLength(20));
-        }
-
-        clusterState = buildClusterState(indexName, indexSettingsBuilder, lifecycleState.build(), policyMetadatas);
-        index = clusterState.metaData().index(indexName).getIndex();
-        newClusterState = IndexLifecycleRunner.moveClusterStateToNextStep(index, clusterState, currentStep, nextStep, () -> now, false);
-        assertClusterStateOnNextStep(clusterState, index, currentStep, nextStep, newClusterState, now);
-    }
-
-    public void testMoveClusterStateToNextStepSamePhase() {
-        String indexName = "my_index";
-        StepKey currentStep = new StepKey("current_phase", "current_action", "current_step");
-        StepKey nextStep = new StepKey("current_phase", "next_action", "next_step");
-        long now = randomNonNegativeLong();
-
-        ClusterState clusterState = buildClusterState(indexName, Settings.builder(), LifecycleExecutionState.builder().build(),
-            Collections.emptyList());
-        Index index = clusterState.metaData().index(indexName).getIndex();
-        ClusterState newClusterState = IndexLifecycleRunner.moveClusterStateToNextStep(index, clusterState, currentStep, nextStep,
-                () -> now, false);
-        assertClusterStateOnNextStep(clusterState, index, currentStep, nextStep, newClusterState, now);
-
-        LifecycleExecutionState.Builder lifecycleState = LifecycleExecutionState.builder();
-        lifecycleState.setPhase(currentStep.getPhase());
-        lifecycleState.setAction(currentStep.getAction());
-        lifecycleState.setStep(currentStep.getName());
-        if (randomBoolean()) {
-            lifecycleState.setStepInfo(randomAlphaOfLength(20));
-        }
-
-        clusterState = buildClusterState(indexName, Settings.builder(), lifecycleState.build(), Collections.emptyList());
-        index = clusterState.metaData().index(indexName).getIndex();
-        newClusterState = IndexLifecycleRunner.moveClusterStateToNextStep(index, clusterState, currentStep, nextStep, () -> now, false);
-        assertClusterStateOnNextStep(clusterState, index, currentStep, nextStep, newClusterState, now);
-    }
-
-    public void testMoveClusterStateToNextStepSameAction() {
-        String indexName = "my_index";
-        StepKey currentStep = new StepKey("current_phase", "current_action", "current_step");
-        StepKey nextStep = new StepKey("current_phase", "current_action", "next_step");
-        long now = randomNonNegativeLong();
-
-        ClusterState clusterState = buildClusterState(indexName, Settings.builder(), LifecycleExecutionState.builder().build(),
-            Collections.emptyList());
-        Index index = clusterState.metaData().index(indexName).getIndex();
-        ClusterState newClusterState = IndexLifecycleRunner.moveClusterStateToNextStep(index, clusterState, currentStep, nextStep,
-                () -> now, false);
-        assertClusterStateOnNextStep(clusterState, index, currentStep, nextStep, newClusterState, now);
-
-        LifecycleExecutionState.Builder lifecycleState = LifecycleExecutionState.builder();
-        lifecycleState.setPhase(currentStep.getPhase());
-        lifecycleState.setAction(currentStep.getAction());
-        lifecycleState.setStep(currentStep.getName());
-        if (randomBoolean()) {
-            lifecycleState.setStepInfo(randomAlphaOfLength(20));
-        }
-        clusterState = buildClusterState(indexName, Settings.builder(), lifecycleState.build(), Collections.emptyList());
-        index = clusterState.metaData().index(indexName).getIndex();
-        newClusterState = IndexLifecycleRunner.moveClusterStateToNextStep(index, clusterState, currentStep, nextStep, () -> now, false);
-        assertClusterStateOnNextStep(clusterState, index, currentStep, nextStep, newClusterState, now);
-    }
-
-    public void testSuccessfulValidatedMoveClusterStateToNextStep() {
-        String indexName = "my_index";
-        String policyName = "my_policy";
-        LifecyclePolicy policy = randomValueOtherThanMany(p -> p.getPhases().size() == 0,
-            () -> LifecyclePolicyTests.randomTestLifecyclePolicy(policyName));
-        Phase nextPhase = policy.getPhases().values().stream().findFirst().get();
-        List<LifecyclePolicyMetadata> policyMetadatas = Collections.singletonList(
-            new LifecyclePolicyMetadata(policy, Collections.emptyMap(), randomNonNegativeLong(), randomNonNegativeLong()));
-        StepKey currentStepKey = new StepKey("current_phase", "current_action", "current_step");
-        StepKey nextStepKey = new StepKey(nextPhase.getName(), "next_action", "next_step");
-        long now = randomNonNegativeLong();
-        Step step = new MockStep(nextStepKey, nextStepKey);
-        PolicyStepsRegistry stepRegistry = createOneStepPolicyStepRegistry(policyName, step, indexName);
-
-        LifecycleExecutionState.Builder lifecycleState = LifecycleExecutionState.builder();
-        lifecycleState.setPhase(currentStepKey.getPhase());
-        lifecycleState.setAction(currentStepKey.getAction());
-        lifecycleState.setStep(currentStepKey.getName());
-
-        Builder indexSettingsBuilder = Settings.builder().put(LifecycleSettings.LIFECYCLE_NAME, policyName);
-        ClusterState clusterState = buildClusterState(indexName, indexSettingsBuilder, lifecycleState.build(), policyMetadatas);
-        Index index = clusterState.metaData().index(indexName).getIndex();
-        ClusterState newClusterState = IndexLifecycleRunner.moveClusterStateToStep(indexName, clusterState, currentStepKey,
-            nextStepKey, () -> now, stepRegistry);
-        assertClusterStateOnNextStep(clusterState, index, currentStepKey, nextStepKey, newClusterState, now);
-    }
-
-    public void testValidatedMoveClusterStateToNextStepWithoutPolicy() {
-        String indexName = "my_index";
-        String policyName = "policy";
-        StepKey currentStepKey = new StepKey("current_phase", "current_action", "current_step");
-        StepKey nextStepKey = new StepKey("next_phase", "next_action", "next_step");
-        long now = randomNonNegativeLong();
-        Step step = new MockStep(nextStepKey, nextStepKey);
-        PolicyStepsRegistry stepRegistry = createOneStepPolicyStepRegistry(policyName, step);
-
-        Builder indexSettingsBuilder = Settings.builder().put(LifecycleSettings.LIFECYCLE_NAME, randomBoolean() ? "" : null);
-        LifecycleExecutionState.Builder lifecycleState = LifecycleExecutionState.builder();
-        lifecycleState.setPhase(currentStepKey.getPhase());
-        lifecycleState.setAction(currentStepKey.getAction());
-        lifecycleState.setStep(currentStepKey.getName());
-
-        ClusterState clusterState = buildClusterState(indexName, indexSettingsBuilder, lifecycleState.build(), Collections.emptyList());
-        IllegalArgumentException exception = expectThrows(IllegalArgumentException.class,
-            () -> IndexLifecycleRunner.moveClusterStateToStep(indexName, clusterState, currentStepKey,
-                nextStepKey, () -> now, stepRegistry));
-        assertThat(exception.getMessage(), equalTo("index [my_index] is not associated with an Index Lifecycle Policy"));
-    }
-
-    public void testValidatedMoveClusterStateToNextStepInvalidCurrentStep() {
-        String indexName = "my_index";
-        String policyName = "my_policy";
-        StepKey currentStepKey = new StepKey("current_phase", "current_action", "current_step");
-        StepKey notCurrentStepKey = new StepKey("not_current_phase", "not_current_action", "not_current_step");
-        StepKey nextStepKey = new StepKey("next_phase", "next_action", "next_step");
-        long now = randomNonNegativeLong();
-        Step step = new MockStep(nextStepKey, nextStepKey);
-        PolicyStepsRegistry stepRegistry = createOneStepPolicyStepRegistry(policyName, step);
-
-        Builder indexSettingsBuilder = Settings.builder().put(LifecycleSettings.LIFECYCLE_NAME, policyName);
-        LifecycleExecutionState.Builder lifecycleState = LifecycleExecutionState.builder();
-        lifecycleState.setPhase(currentStepKey.getPhase());
-        lifecycleState.setAction(currentStepKey.getAction());
-        lifecycleState.setStep(currentStepKey.getName());
-
-        ClusterState clusterState = buildClusterState(indexName, indexSettingsBuilder, lifecycleState.build(), Collections.emptyList());
-        IllegalArgumentException exception = expectThrows(IllegalArgumentException.class,
-            () -> IndexLifecycleRunner.moveClusterStateToStep(indexName, clusterState, notCurrentStepKey,
-                nextStepKey, () -> now, stepRegistry));
-        assertThat(exception.getMessage(), equalTo("index [my_index] is not on current step " +
-            "[{\"phase\":\"not_current_phase\",\"action\":\"not_current_action\",\"name\":\"not_current_step\"}]"));
-    }
-
-    public void testValidatedMoveClusterStateToNextStepInvalidNextStep() {
-        String indexName = "my_index";
-        String policyName = "my_policy";
-        StepKey currentStepKey = new StepKey("current_phase", "current_action", "current_step");
-        StepKey nextStepKey = new StepKey("next_phase", "next_action", "next_step");
-        long now = randomNonNegativeLong();
-        Step step = new MockStep(currentStepKey, nextStepKey);
-        PolicyStepsRegistry stepRegistry = createOneStepPolicyStepRegistry(policyName, step);
-
-        Builder indexSettingsBuilder = Settings.builder().put(LifecycleSettings.LIFECYCLE_NAME, policyName);
-        LifecycleExecutionState.Builder lifecycleState = LifecycleExecutionState.builder();
-        lifecycleState.setPhase(currentStepKey.getPhase());
-        lifecycleState.setAction(currentStepKey.getAction());
-        lifecycleState.setStep(currentStepKey.getName());
-
-        ClusterState clusterState = buildClusterState(indexName, indexSettingsBuilder, lifecycleState.build(), Collections.emptyList());
-        IllegalArgumentException exception = expectThrows(IllegalArgumentException.class,
-            () -> IndexLifecycleRunner.moveClusterStateToStep(indexName, clusterState, currentStepKey,
-                nextStepKey, () -> now, stepRegistry));
-        assertThat(exception.getMessage(),
-            equalTo("step [{\"phase\":\"next_phase\",\"action\":\"next_action\",\"name\":\"next_step\"}] " +
-                "for index [my_index] with policy [my_policy] does not exist"));
-    }
-
-    public void testMoveClusterStateToErrorStep() throws IOException {
-        String indexName = "my_index";
-        StepKey currentStep = new StepKey("current_phase", "current_action", "current_step");
-        StepKey nextStepKey = new StepKey("next_phase", "next_action", "next_step");
-        long now = randomNonNegativeLong();
-        Exception cause = new ElasticsearchException("THIS IS AN EXPECTED CAUSE");
-
-        LifecycleExecutionState.Builder lifecycleState = LifecycleExecutionState.builder();
-        lifecycleState.setPhase(currentStep.getPhase());
-        lifecycleState.setAction(currentStep.getAction());
-        lifecycleState.setStep(currentStep.getName());
-        ClusterState clusterState = buildClusterState(indexName, Settings.builder(), lifecycleState.build(), Collections.emptyList());
-        Index index = clusterState.metaData().index(indexName).getIndex();
-
-        ClusterState newClusterState = IndexLifecycleRunner.moveClusterStateToErrorStep(index, clusterState, currentStep, cause, () -> now,
-            (idxMeta, stepKey) -> new MockStep(stepKey, nextStepKey));
-        assertClusterStateOnErrorStep(clusterState, index, currentStep, newClusterState, now,
-            "{\"type\":\"exception\",\"reason\":\"THIS IS AN EXPECTED CAUSE\"");
-
-        cause = new IllegalArgumentException("non elasticsearch-exception");
-        newClusterState = IndexLifecycleRunner.moveClusterStateToErrorStep(index, clusterState, currentStep, cause, () -> now,
-            (idxMeta, stepKey) -> new MockStep(stepKey, nextStepKey));
-        assertClusterStateOnErrorStep(clusterState, index, currentStep, newClusterState, now,
-            "{\"type\":\"illegal_argument_exception\",\"reason\":\"non elasticsearch-exception\",\"stack_trace\":\"");
-    }
-
-    public void testMoveClusterStateToFailedStep() {
-        String indexName = "my_index";
-        String[] indices = new String[] { indexName };
-        String policyName = "my_policy";
-        long now = randomNonNegativeLong();
-        StepKey failedStepKey = new StepKey("current_phase", MockAction.NAME, "current_step");
-        StepKey errorStepKey = new StepKey(failedStepKey.getPhase(), failedStepKey.getAction(), ErrorStep.NAME);
-        Step step = new MockStep(failedStepKey, null);
-        LifecyclePolicy policy = createPolicy(policyName, failedStepKey, null);
-        LifecyclePolicyMetadata policyMetadata = new LifecyclePolicyMetadata(policy, Collections.emptyMap(),
-            randomNonNegativeLong(), randomNonNegativeLong());
-
-        PolicyStepsRegistry policyRegistry = createOneStepPolicyStepRegistry(policyName, step, indexName);
-        Settings.Builder indexSettingsBuilder = Settings.builder()
-                .put(LifecycleSettings.LIFECYCLE_NAME, policyName);
-        LifecycleExecutionState.Builder lifecycleState = LifecycleExecutionState.builder();
-        lifecycleState.setPhase(errorStepKey.getPhase());
-        lifecycleState.setPhaseTime(now);
-        lifecycleState.setAction(errorStepKey.getAction());
-        lifecycleState.setActionTime(now);
-        lifecycleState.setStep(errorStepKey.getName());
-        lifecycleState.setStepTime(now);
-        lifecycleState.setFailedStep(failedStepKey.getName());
-        ClusterState clusterState = buildClusterState(indexName, indexSettingsBuilder, lifecycleState.build(),
-            Collections.singletonList(policyMetadata));
-        Index index = clusterState.metaData().index(indexName).getIndex();
-        IndexLifecycleRunner runner = new IndexLifecycleRunner(policyRegistry, null, threadPool, () -> now);
-        ClusterState nextClusterState = runner.moveClusterStateToPreviouslyFailedStep(clusterState, indices);
-        IndexLifecycleRunnerTests.assertClusterStateOnNextStep(clusterState, index, errorStepKey, failedStepKey,
-            nextClusterState, now);
-        LifecycleExecutionState executionState = LifecycleExecutionState.fromIndexMetadata(nextClusterState.metaData().index(indexName));
-        assertThat("manual move to failed step should not count as a retry", executionState.getFailedStepRetryCount(), is(nullValue()));
-    }
-
-    public void testMoveClusterStateToFailedStepWithUnknownStep() {
-        String indexName = "my_index";
-        String[] indices = new String[] { indexName };
-        String policyName = "my_policy";
-        long now = randomNonNegativeLong();
-        StepKey failedStepKey = new StepKey("current_phase", MockAction.NAME, "current_step");
-        StepKey errorStepKey = new StepKey(failedStepKey.getPhase(), failedStepKey.getAction(), ErrorStep.NAME);
-
-        StepKey registeredStepKey = new StepKey(randomFrom(failedStepKey.getPhase(), "other"),
-            MockAction.NAME, "different_step");
-        Step step = new MockStep(registeredStepKey, null);
-        LifecyclePolicy policy = createPolicy(policyName, failedStepKey, null);
-        LifecyclePolicyMetadata policyMetadata = new LifecyclePolicyMetadata(policy, Collections.emptyMap(),
-            randomNonNegativeLong(), randomNonNegativeLong());
-
-        PolicyStepsRegistry policyRegistry = createOneStepPolicyStepRegistry(policyName, step, indexName);
-        Settings.Builder indexSettingsBuilder = Settings.builder()
-            .put(LifecycleSettings.LIFECYCLE_NAME, policyName);
-        LifecycleExecutionState.Builder lifecycleState = LifecycleExecutionState.builder();
-        lifecycleState.setPhase(errorStepKey.getPhase());
-        lifecycleState.setPhaseTime(now);
-        lifecycleState.setAction(errorStepKey.getAction());
-        lifecycleState.setActionTime(now);
-        lifecycleState.setStep(errorStepKey.getName());
-        lifecycleState.setStepTime(now);
-        lifecycleState.setFailedStep(failedStepKey.getName());
-        ClusterState clusterState = buildClusterState(indexName, indexSettingsBuilder, lifecycleState.build(),
-            Collections.singletonList(policyMetadata));
-        IndexLifecycleRunner runner = new IndexLifecycleRunner(policyRegistry, null, threadPool, () -> now);
-        IllegalArgumentException exception = expectThrows(IllegalArgumentException.class,
-            () -> runner.moveClusterStateToPreviouslyFailedStep(clusterState, indices));
-        assertThat(exception.getMessage(), equalTo("step [" + failedStepKey
-            + "] for index [my_index] with policy [my_policy] does not exist"));
-    }
-
-    public void testMoveClusterStateToFailedStepIndexNotFound() {
-        String existingIndexName = "my_index";
-        String invalidIndexName = "does_not_exist";
-        ClusterState clusterState = buildClusterState(existingIndexName, Settings.builder(), LifecycleExecutionState.builder().build(),
-            Collections.emptyList());
-        IndexLifecycleRunner runner = new IndexLifecycleRunner(null, null, threadPool, () -> 0L);
-        IllegalArgumentException exception = expectThrows(IllegalArgumentException.class,
-            () -> runner.moveClusterStateToPreviouslyFailedStep(clusterState, new String[] { invalidIndexName }));
-        assertThat(exception.getMessage(), equalTo("index [" + invalidIndexName + "] does not exist"));
-    }
-
-    public void testMoveClusterStateToFailedStepInvalidPolicySetting() {
-        String indexName = "my_index";
-        String[] indices = new String[] { indexName };
-        String policyName = "my_policy";
-        long now = randomNonNegativeLong();
-        StepKey failedStepKey = new StepKey("current_phase", "current_action", "current_step");
-        StepKey errorStepKey = new StepKey(failedStepKey.getPhase(), failedStepKey.getAction(), ErrorStep.NAME);
-        Step step = new MockStep(failedStepKey, null);
-        PolicyStepsRegistry policyRegistry = createOneStepPolicyStepRegistry(policyName, step);
-        Settings.Builder indexSettingsBuilder = Settings.builder()
-            .put(LifecycleSettings.LIFECYCLE_NAME, (String) null);
-        LifecycleExecutionState.Builder lifecycleState = LifecycleExecutionState.builder();
-        lifecycleState.setPhase(errorStepKey.getPhase());
-        lifecycleState.setAction(errorStepKey.getAction());
-        lifecycleState.setStep(errorStepKey.getName());
-        lifecycleState.setFailedStep(failedStepKey.getName());
-        ClusterState clusterState = buildClusterState(indexName, indexSettingsBuilder, lifecycleState.build(), Collections.emptyList());
-        IndexLifecycleRunner runner = new IndexLifecycleRunner(policyRegistry, null, threadPool, () -> now);
-        IllegalArgumentException exception = expectThrows(IllegalArgumentException.class,
-            () -> runner.moveClusterStateToPreviouslyFailedStep(clusterState, indices));
-        assertThat(exception.getMessage(), equalTo("index [" + indexName + "] is not associated with an Index Lifecycle Policy"));
-    }
-
-    public void testMoveClusterStateToFailedNotOnError() {
-        String indexName = "my_index";
-        String[] indices = new String[] { indexName };
-        String policyName = "my_policy";
-        long now = randomNonNegativeLong();
-        StepKey failedStepKey = new StepKey("current_phase", "current_action", "current_step");
-        Step step = new MockStep(failedStepKey, null);
-        PolicyStepsRegistry policyRegistry = createOneStepPolicyStepRegistry(policyName, step);
-        Settings.Builder indexSettingsBuilder = Settings.builder()
-            .put(LifecycleSettings.LIFECYCLE_NAME, (String) null);
-        LifecycleExecutionState.Builder lifecycleState = LifecycleExecutionState.builder();
-        lifecycleState.setPhase(failedStepKey.getPhase());
-        lifecycleState.setAction(failedStepKey.getAction());
-        lifecycleState.setStep(failedStepKey.getName());
-        ClusterState clusterState = buildClusterState(indexName, indexSettingsBuilder, lifecycleState.build(), Collections.emptyList());
-        IndexLifecycleRunner runner = new IndexLifecycleRunner(policyRegistry, null, threadPool, () -> now);
-        IllegalArgumentException exception = expectThrows(IllegalArgumentException.class,
-            () -> runner.moveClusterStateToPreviouslyFailedStep(clusterState, indices));
-        assertThat(exception.getMessage(), equalTo("cannot retry an action for an index [" + indices[0]
-            + "] that has not encountered an error when running a Lifecycle Policy"));
-    }
-
-    public void testMoveClusterStateToPreviouslyFailedStepAsAutomaticRetry() {
-        String indexName = "my_index";
-        String policyName = "my_policy";
-        long now = randomNonNegativeLong();
-        StepKey failedStepKey = new StepKey("current_phase", MockAction.NAME, "current_step");
-        StepKey errorStepKey = new StepKey(failedStepKey.getPhase(), failedStepKey.getAction(), ErrorStep.NAME);
-        Step retryableStep = new RetryableMockStep(failedStepKey, null);
-        LifecyclePolicy policy = createPolicy(policyName, failedStepKey, null);
-        LifecyclePolicyMetadata policyMetadata = new LifecyclePolicyMetadata(policy, Collections.emptyMap(),
-            randomNonNegativeLong(), randomNonNegativeLong());
-
-        PolicyStepsRegistry policyRegistry = createOneStepPolicyStepRegistry(policyName, retryableStep, indexName);
-        Settings.Builder indexSettingsBuilder = Settings.builder()
-                .put(LifecycleSettings.LIFECYCLE_NAME, policyName);
-        LifecycleExecutionState.Builder lifecycleState = LifecycleExecutionState.builder();
-        lifecycleState.setPhase(errorStepKey.getPhase());
-        lifecycleState.setPhaseTime(now);
-        lifecycleState.setAction(errorStepKey.getAction());
-        lifecycleState.setActionTime(now);
-        lifecycleState.setStep(errorStepKey.getName());
-        lifecycleState.setStepTime(now);
-        lifecycleState.setFailedStep(failedStepKey.getName());
-        ClusterState clusterState = buildClusterState(indexName, indexSettingsBuilder, lifecycleState.build(),
-            Collections.singletonList(policyMetadata));
-        Index index = clusterState.metaData().index(indexName).getIndex();
-        IndexLifecycleRunner runner = new IndexLifecycleRunner(policyRegistry, null, threadPool, () -> now);
-        ClusterState nextClusterState = runner.moveClusterStateToPreviouslyFailedStep(clusterState, indexName, true);
-        IndexLifecycleRunnerTests.assertClusterStateOnNextStep(clusterState, index, errorStepKey, failedStepKey,
-            nextClusterState, now);
-        LifecycleExecutionState executionState = LifecycleExecutionState.fromIndexMetadata(nextClusterState.metaData().index(indexName));
-        assertThat(executionState.getFailedStepRetryCount(), is(1));
-    }
-
-    public void testAddStepInfoToClusterState() throws IOException {
-        String indexName = "my_index";
-        StepKey currentStep = new StepKey("current_phase", "current_action", "current_step");
-        RandomStepInfo stepInfo = new RandomStepInfo(() -> randomAlphaOfLength(10));
-
-        LifecycleExecutionState.Builder lifecycleState = LifecycleExecutionState.builder();
-        lifecycleState.setPhase(currentStep.getPhase());
-        lifecycleState.setAction(currentStep.getAction());
-        lifecycleState.setStep(currentStep.getName());
-        ClusterState clusterState = buildClusterState(indexName, Settings.builder(), lifecycleState.build(), Collections.emptyList());
-        Index index = clusterState.metaData().index(indexName).getIndex();
-        ClusterState newClusterState = IndexLifecycleRunner.addStepInfoToClusterState(index, clusterState, stepInfo);
-        assertClusterStateStepInfo(clusterState, index, currentStep, newClusterState, stepInfo);
-        ClusterState runAgainClusterState = IndexLifecycleRunner.addStepInfoToClusterState(index, newClusterState, stepInfo);
-        assertSame(newClusterState, runAgainClusterState);
-    }
-
-    private ClusterState buildClusterState(String indexName, Settings.Builder indexSettingsBuilder,
-                                           LifecycleExecutionState lifecycleState,
-                                           List<LifecyclePolicyMetadata> lifecyclePolicyMetadatas) {
-        Settings indexSettings = indexSettingsBuilder.put(IndexMetaData.SETTING_NUMBER_OF_SHARDS, 1)
-                .put(IndexMetaData.SETTING_NUMBER_OF_REPLICAS, 0).put(IndexMetaData.SETTING_VERSION_CREATED, Version.CURRENT).build();
-        IndexMetaData indexMetadata = IndexMetaData.builder(indexName)
-            .settings(indexSettings)
-            .putCustom(ILM_CUSTOM_METADATA_KEY, lifecycleState.asMap())
-            .build();
-
-        Map<String, LifecyclePolicyMetadata> lifecyclePolicyMetadatasMap = lifecyclePolicyMetadatas.stream()
-                .collect(Collectors.toMap(LifecyclePolicyMetadata::getName, Function.identity()));
-        IndexLifecycleMetadata indexLifecycleMetadata = new IndexLifecycleMetadata(lifecyclePolicyMetadatasMap, OperationMode.RUNNING);
-
-        MetaData metadata = MetaData.builder().put(indexMetadata, true).putCustom(IndexLifecycleMetadata.TYPE, indexLifecycleMetadata)
-                .build();
-        return ClusterState.builder(new ClusterName("my_cluster")).metaData(metadata).build();
-    }
-
-    private static LifecyclePolicy createPolicy(String policyName, StepKey safeStep, StepKey unsafeStep) {
-        Map<String, Phase> phases = new HashMap<>();
-        if (safeStep != null) {
-            assert MockAction.NAME.equals(safeStep.getAction()) : "The safe action needs to be MockAction.NAME";
-            assert unsafeStep == null
-                    || safeStep.getPhase().equals(unsafeStep.getPhase()) == false : "safe and unsafe actions must be in different phases";
-            Map<String, LifecycleAction> actions = new HashMap<>();
-            List<Step> steps = Collections.singletonList(new MockStep(safeStep, null));
-            MockAction safeAction = new MockAction(steps, true);
-            actions.put(safeAction.getWriteableName(), safeAction);
-            Phase phase = new Phase(safeStep.getPhase(), TimeValue.timeValueMillis(0), actions);
-            phases.put(phase.getName(), phase);
-        }
-        if (unsafeStep != null) {
-            assert MockAction.NAME.equals(unsafeStep.getAction()) : "The unsafe action needs to be MockAction.NAME";
-            Map<String, LifecycleAction> actions = new HashMap<>();
-            List<Step> steps = Collections.singletonList(new MockStep(unsafeStep, null));
-            MockAction unsafeAction = new MockAction(steps, false);
-            actions.put(unsafeAction.getWriteableName(), unsafeAction);
-            Phase phase = new Phase(unsafeStep.getPhase(), TimeValue.timeValueMillis(0), actions);
-            phases.put(phase.getName(), phase);
-        }
-        return newTestLifecyclePolicy(policyName, phases);
-    }
-
-    public void testRemovePolicyForIndex() {
-        String indexName = randomAlphaOfLength(10);
-        String oldPolicyName = "old_policy";
-        StepKey currentStep = new StepKey(randomAlphaOfLength(10), MockAction.NAME, randomAlphaOfLength(10));
-        LifecyclePolicy oldPolicy = createPolicy(oldPolicyName, currentStep, null);
-        Settings.Builder indexSettingsBuilder = Settings.builder().put(LifecycleSettings.LIFECYCLE_NAME, oldPolicyName);
-        LifecycleExecutionState.Builder lifecycleState = LifecycleExecutionState.builder();
-        lifecycleState.setPhase(currentStep.getPhase());
-        lifecycleState.setAction(currentStep.getAction());
-        lifecycleState.setStep(currentStep.getName());
-        List<LifecyclePolicyMetadata> policyMetadatas = new ArrayList<>();
-        policyMetadatas.add(new LifecyclePolicyMetadata(oldPolicy, Collections.emptyMap(),
-            randomNonNegativeLong(), randomNonNegativeLong()));
-        ClusterState clusterState = buildClusterState(indexName, indexSettingsBuilder, lifecycleState.build(), policyMetadatas);
-        Index index = clusterState.metaData().index(indexName).getIndex();
-        Index[] indices = new Index[] { index };
-        List<String> failedIndexes = new ArrayList<>();
-
-        ClusterState newClusterState = IndexLifecycleRunner.removePolicyForIndexes(indices, clusterState, failedIndexes);
-
-        assertTrue(failedIndexes.isEmpty());
-        assertIndexNotManagedByILM(newClusterState, index);
-    }
-
-    public void testRemovePolicyForIndexNoCurrentPolicy() {
-        String indexName = randomAlphaOfLength(10);
-        Settings.Builder indexSettingsBuilder = Settings.builder();
-        ClusterState clusterState = buildClusterState(indexName, indexSettingsBuilder, LifecycleExecutionState.builder().build(),
-            Collections.emptyList());
-        Index index = clusterState.metaData().index(indexName).getIndex();
-        Index[] indices = new Index[] { index };
-        List<String> failedIndexes = new ArrayList<>();
-
-        ClusterState newClusterState = IndexLifecycleRunner.removePolicyForIndexes(indices, clusterState, failedIndexes);
-
-        assertTrue(failedIndexes.isEmpty());
-        assertIndexNotManagedByILM(newClusterState, index);
-    }
-
-    public void testRemovePolicyForIndexIndexDoesntExist() {
-        String indexName = randomAlphaOfLength(10);
-        String oldPolicyName = "old_policy";
-        LifecyclePolicy oldPolicy = newTestLifecyclePolicy(oldPolicyName, Collections.emptyMap());
-        StepKey currentStep = AbstractStepTestCase.randomStepKey();
-        Settings.Builder indexSettingsBuilder = Settings.builder().put(LifecycleSettings.LIFECYCLE_NAME, oldPolicyName);
-        LifecycleExecutionState.Builder lifecycleState = LifecycleExecutionState.builder();
-        lifecycleState.setPhase(currentStep.getPhase());
-        lifecycleState.setAction(currentStep.getAction());
-        lifecycleState.setStep(currentStep.getName());
-        List<LifecyclePolicyMetadata> policyMetadatas = new ArrayList<>();
-        policyMetadatas.add(new LifecyclePolicyMetadata(oldPolicy, Collections.emptyMap(),
-            randomNonNegativeLong(), randomNonNegativeLong()));
-        ClusterState clusterState = buildClusterState(indexName, indexSettingsBuilder, lifecycleState.build(), policyMetadatas);
-        Index index = new Index("doesnt_exist", "im_not_here");
-        Index[] indices = new Index[] { index };
-        List<String> failedIndexes = new ArrayList<>();
-
-        ClusterState newClusterState = IndexLifecycleRunner.removePolicyForIndexes(indices, clusterState, failedIndexes);
-
-        assertEquals(1, failedIndexes.size());
-        assertEquals("doesnt_exist", failedIndexes.get(0));
-        assertSame(clusterState, newClusterState);
-    }
-
-    public void testRemovePolicyForIndexIndexInUnsafe() {
-        String indexName = randomAlphaOfLength(10);
-        String oldPolicyName = "old_policy";
-        StepKey currentStep = new StepKey(randomAlphaOfLength(10), MockAction.NAME, randomAlphaOfLength(10));
-        LifecyclePolicy oldPolicy = createPolicy(oldPolicyName, null, currentStep);
-        Settings.Builder indexSettingsBuilder = Settings.builder().put(LifecycleSettings.LIFECYCLE_NAME, oldPolicyName);
-        LifecycleExecutionState.Builder lifecycleState = LifecycleExecutionState.builder();
-        lifecycleState.setPhase(currentStep.getPhase());
-        lifecycleState.setAction(currentStep.getAction());
-        lifecycleState.setStep(currentStep.getName());
-        List<LifecyclePolicyMetadata> policyMetadatas = new ArrayList<>();
-        policyMetadatas.add(new LifecyclePolicyMetadata(oldPolicy, Collections.emptyMap(),
-            randomNonNegativeLong(), randomNonNegativeLong()));
-        ClusterState clusterState = buildClusterState(indexName, indexSettingsBuilder, lifecycleState.build(), policyMetadatas);
-        Index index = clusterState.metaData().index(indexName).getIndex();
-        Index[] indices = new Index[] { index };
-        List<String> failedIndexes = new ArrayList<>();
-
-        ClusterState newClusterState = IndexLifecycleRunner.removePolicyForIndexes(indices, clusterState, failedIndexes);
-
-        assertTrue(failedIndexes.isEmpty());
-        assertIndexNotManagedByILM(newClusterState, index);
-    }
-
-    public void testRemovePolicyWithIndexingComplete() {
-        String indexName = randomAlphaOfLength(10);
-        String oldPolicyName = "old_policy";
-        StepKey currentStep = new StepKey(randomAlphaOfLength(10), MockAction.NAME, randomAlphaOfLength(10));
-        LifecyclePolicy oldPolicy = createPolicy(oldPolicyName, null, currentStep);
-        Settings.Builder indexSettingsBuilder = Settings.builder()
-            .put(LifecycleSettings.LIFECYCLE_NAME, oldPolicyName)
-            .put(LifecycleSettings.LIFECYCLE_INDEXING_COMPLETE, true);
-        LifecycleExecutionState.Builder lifecycleState = LifecycleExecutionState.builder();
-        lifecycleState.setPhase(currentStep.getPhase());
-        lifecycleState.setAction(currentStep.getAction());
-        lifecycleState.setStep(currentStep.getName());
-        List<LifecyclePolicyMetadata> policyMetadatas = new ArrayList<>();
-        policyMetadatas.add(new LifecyclePolicyMetadata(oldPolicy, Collections.emptyMap(),
-            randomNonNegativeLong(), randomNonNegativeLong()));
-        ClusterState clusterState = buildClusterState(indexName, indexSettingsBuilder, lifecycleState.build(), policyMetadatas);
-        Index index = clusterState.metaData().index(indexName).getIndex();
-        Index[] indices = new Index[] { index };
-        List<String> failedIndexes = new ArrayList<>();
-
-        ClusterState newClusterState = IndexLifecycleRunner.removePolicyForIndexes(indices, clusterState, failedIndexes);
-
-        assertTrue(failedIndexes.isEmpty());
-        assertIndexNotManagedByILM(newClusterState, index);
     }
 
     public void testIsReadyToTransition() {
@@ -1405,120 +751,29 @@ public class IndexLifecycleRunnerTests extends ESTestCase {
             runner.isReadyToTransitionToThisPhase(policyName, indexMetaData, "phase"));
     }
 
-   public void testValidateTransitionThrowsExceptionForMissingIndexPolicy() {
-       IndexMetaData indexMetaData = IndexMetaData.builder("index").settings(settings(Version.CURRENT))
-           .numberOfShards(randomIntBetween(1, 5))
-           .numberOfReplicas(randomIntBetween(0, 5))
-           .build();
-
-       StepKey currentStepKey = new StepKey("hot", "action", "firstStep");
-       StepKey nextStepKey = new StepKey("hot", "action", "secondStep");
-       Step currentStep = new MockStep(currentStepKey, nextStepKey);
-       MockPolicyStepsRegistry policyRegistry = createOneStepPolicyStepRegistry("policy", currentStep);
-
-       expectThrows(IllegalArgumentException.class,
-           () -> IndexLifecycleRunner.validateTransition(indexMetaData, currentStepKey, nextStepKey, policyRegistry));
-   }
-
-   public void testValidateTransitionThrowsExceptionIfTheCurrentStepIsIncorrect() {
-       LifecycleExecutionState.Builder lifecycleState = LifecycleExecutionState.builder();
-       lifecycleState.setPhase("hot");
-       lifecycleState.setAction("action");
-       lifecycleState.setStep("another_step");
-       String policy = "policy";
-       IndexMetaData indexMetaData = buildIndexMetadata(policy, lifecycleState);
-
-       StepKey currentStepKey = new StepKey("hot", "action", "firstStep");
-       StepKey nextStepKey = new StepKey("hot", "action", "secondStep");
-       Step currentStep = new MockStep(currentStepKey, nextStepKey);
-       MockPolicyStepsRegistry policyRegistry = createOneStepPolicyStepRegistry(policy, currentStep);
-
-       expectThrows(IllegalArgumentException.class,
-           () -> IndexLifecycleRunner.validateTransition(indexMetaData, currentStepKey, nextStepKey, policyRegistry));
-   }
-
-   public void testValidateTransitionThrowsExceptionIfNextStepDoesNotExist() {
-       LifecycleExecutionState.Builder lifecycleState = LifecycleExecutionState.builder();
-       lifecycleState.setPhase("hot");
-       lifecycleState.setAction("action");
-       lifecycleState.setStep("firstStep");
-       String policy = "policy";
-       IndexMetaData indexMetaData = buildIndexMetadata(policy, lifecycleState);
-
-       StepKey currentStepKey = new StepKey("hot", "action", "firstStep");
-       StepKey nextStepKey = new StepKey("hot", "action", "secondStep");
-       Step currentStep = new MockStep(currentStepKey, nextStepKey);
-       MockPolicyStepsRegistry policyRegistry = createOneStepPolicyStepRegistry(policy, currentStep);
-
-       expectThrows(IllegalArgumentException.class,
-           () -> IndexLifecycleRunner.validateTransition(indexMetaData, currentStepKey, nextStepKey, policyRegistry));
-   }
-
-    public void testValidateValidTransition() {
-        LifecycleExecutionState.Builder lifecycleState = LifecycleExecutionState.builder();
-        lifecycleState.setPhase("hot");
-        lifecycleState.setAction("action");
-        lifecycleState.setStep("firstStep");
-        String policy = "policy";
-        IndexMetaData indexMetaData = buildIndexMetadata(policy, lifecycleState);
-
-        StepKey currentStepKey = new StepKey("hot", "action", "firstStep");
-        StepKey nextStepKey = new StepKey("hot", "action", "secondStep");
-        Step finalStep = new MockStep(nextStepKey, new StepKey("hot", "action", "completed"));
-        MockPolicyStepsRegistry policyRegistry = createOneStepPolicyStepRegistry(policy, finalStep);
-
-        try {
-            IndexLifecycleRunner.validateTransition(indexMetaData, currentStepKey, nextStepKey, policyRegistry);
-        } catch (Exception e) {
-            logger.error(e);
-            fail("validateTransition should not throw exception on valid transitions");
+    private static LifecyclePolicy createPolicy(String policyName, StepKey safeStep, StepKey unsafeStep) {
+        Map<String, Phase> phases = new HashMap<>();
+        if (safeStep != null) {
+            assert MockAction.NAME.equals(safeStep.getAction()) : "The safe action needs to be MockAction.NAME";
+            assert unsafeStep == null
+                || safeStep.getPhase().equals(unsafeStep.getPhase()) == false : "safe and unsafe actions must be in different phases";
+            Map<String, LifecycleAction> actions = new HashMap<>();
+            List<Step> steps = Collections.singletonList(new MockStep(safeStep, null));
+            MockAction safeAction = new MockAction(steps, true);
+            actions.put(safeAction.getWriteableName(), safeAction);
+            Phase phase = new Phase(safeStep.getPhase(), TimeValue.timeValueMillis(0), actions);
+            phases.put(phase.getName(), phase);
         }
-    }
-
-    public static void assertIndexNotManagedByILM(ClusterState clusterState, Index index) {
-        MetaData metadata = clusterState.metaData();
-        assertNotNull(metadata);
-        IndexMetaData indexMetadata = metadata.getIndexSafe(index);
-        assertNotNull(indexMetadata);
-        Settings indexSettings = indexMetadata.getSettings();
-        assertNotNull(indexSettings);
-        assertFalse(LifecycleSettings.LIFECYCLE_NAME_SETTING.exists(indexSettings));
-        assertFalse(RolloverAction.LIFECYCLE_ROLLOVER_ALIAS_SETTING.exists(indexSettings));
-        assertFalse(LifecycleSettings.LIFECYCLE_INDEXING_COMPLETE_SETTING.exists(indexSettings));
-    }
-
-    public static void assertClusterStateOnPolicy(ClusterState oldClusterState, Index index, String expectedPolicy, StepKey previousStep,
-            StepKey expectedStep, ClusterState newClusterState, long now) {
-        assertNotSame(oldClusterState, newClusterState);
-        MetaData newMetadata = newClusterState.metaData();
-        assertNotSame(oldClusterState.metaData(), newMetadata);
-        IndexMetaData newIndexMetadata = newMetadata.getIndexSafe(index);
-        assertNotSame(oldClusterState.metaData().index(index), newIndexMetadata);
-        LifecycleExecutionState newLifecycleState = LifecycleExecutionState
-            .fromIndexMetadata(newClusterState.metaData().index(index));
-        LifecycleExecutionState oldLifecycleState = LifecycleExecutionState
-            .fromIndexMetadata(oldClusterState.metaData().index(index));
-        assertNotSame(oldLifecycleState, newLifecycleState);
-        assertEquals(expectedStep.getPhase(), newLifecycleState.getPhase());
-        assertEquals(expectedStep.getAction(), newLifecycleState.getAction());
-        assertEquals(expectedStep.getName(), newLifecycleState.getStep());
-        if (Objects.equals(previousStep.getPhase(), expectedStep.getPhase())) {
-            assertEquals(oldLifecycleState.getPhase(), newLifecycleState.getPhase());
-        } else {
-            assertEquals(now, newLifecycleState.getPhaseTime().longValue());
+        if (unsafeStep != null) {
+            assert MockAction.NAME.equals(unsafeStep.getAction()) : "The unsafe action needs to be MockAction.NAME";
+            Map<String, LifecycleAction> actions = new HashMap<>();
+            List<Step> steps = Collections.singletonList(new MockStep(unsafeStep, null));
+            MockAction unsafeAction = new MockAction(steps, false);
+            actions.put(unsafeAction.getWriteableName(), unsafeAction);
+            Phase phase = new Phase(unsafeStep.getPhase(), TimeValue.timeValueMillis(0), actions);
+            phases.put(phase.getName(), phase);
         }
-        if (Objects.equals(previousStep.getAction(), expectedStep.getAction())) {
-            assertEquals(oldLifecycleState.getActionTime(), newLifecycleState.getActionTime());
-        } else {
-            assertEquals(now, newLifecycleState.getActionTime().longValue());
-        }
-        if (Objects.equals(previousStep.getName(), expectedStep.getName())) {
-            assertEquals(oldLifecycleState.getStepTime(), newLifecycleState.getStepTime());
-        } else {
-            assertEquals(now, newLifecycleState.getStepTime().longValue());
-        }
-        assertEquals(null, newLifecycleState.getFailedStep());
-        assertEquals(null, newLifecycleState.getStepInfo());
+        return newTestLifecyclePolicy(policyName, phases);
     }
 
     public static void assertClusterStateOnNextStep(ClusterState oldClusterState, Index index, StepKey currentStep, StepKey nextStep,
@@ -1549,61 +804,6 @@ public class IndexLifecycleRunnerTests extends ESTestCase {
         assertEquals(now, newLifecycleState.getStepTime().longValue());
         assertEquals(null, newLifecycleState.getFailedStep());
         assertEquals(null, newLifecycleState.getStepInfo());
-    }
-
-    private IndexMetaData buildIndexMetadata(String policy, LifecycleExecutionState.Builder lifecycleState) {
-        return IndexMetaData.builder("index")
-            .settings(settings(Version.CURRENT).put(LifecycleSettings.LIFECYCLE_NAME, policy))
-            .numberOfShards(randomIntBetween(1, 5))
-            .numberOfReplicas(randomIntBetween(0, 5))
-            .putCustom(ILM_CUSTOM_METADATA_KEY, lifecycleState.build().asMap())
-            .build();
-    }
-
-    private void assertClusterStateOnErrorStep(ClusterState oldClusterState, Index index, StepKey currentStep,
-                                               ClusterState newClusterState, long now, String expectedCauseValue) throws IOException {
-        assertNotSame(oldClusterState, newClusterState);
-        MetaData newMetadata = newClusterState.metaData();
-        assertNotSame(oldClusterState.metaData(), newMetadata);
-        IndexMetaData newIndexMetadata = newMetadata.getIndexSafe(index);
-        assertNotSame(oldClusterState.metaData().index(index), newIndexMetadata);
-        LifecycleExecutionState newLifecycleState = LifecycleExecutionState
-            .fromIndexMetadata(newClusterState.metaData().index(index));
-        LifecycleExecutionState oldLifecycleState = LifecycleExecutionState
-            .fromIndexMetadata(oldClusterState.metaData().index(index));
-        assertNotSame(oldLifecycleState, newLifecycleState);
-        assertEquals(currentStep.getPhase(), newLifecycleState.getPhase());
-        assertEquals(currentStep.getAction(), newLifecycleState.getAction());
-        assertEquals(ErrorStep.NAME, newLifecycleState.getStep());
-        assertEquals(currentStep.getName(), newLifecycleState.getFailedStep());
-        assertThat(newLifecycleState.getStepInfo(), containsString(expectedCauseValue));
-        assertEquals(oldLifecycleState.getPhaseTime(), newLifecycleState.getPhaseTime());
-        assertEquals(oldLifecycleState.getActionTime(), newLifecycleState.getActionTime());
-        assertEquals(now, newLifecycleState.getStepTime().longValue());
-    }
-
-    private void assertClusterStateStepInfo(ClusterState oldClusterState, Index index, StepKey currentStep, ClusterState newClusterState,
-            ToXContentObject stepInfo) throws IOException {
-        XContentBuilder stepInfoXContentBuilder = JsonXContent.contentBuilder();
-        stepInfo.toXContent(stepInfoXContentBuilder, ToXContent.EMPTY_PARAMS);
-        String expectedstepInfoValue = BytesReference.bytes(stepInfoXContentBuilder).utf8ToString();
-        assertNotSame(oldClusterState, newClusterState);
-        MetaData newMetadata = newClusterState.metaData();
-        assertNotSame(oldClusterState.metaData(), newMetadata);
-        IndexMetaData newIndexMetadata = newMetadata.getIndexSafe(index);
-        assertNotSame(oldClusterState.metaData().index(index), newIndexMetadata);
-        LifecycleExecutionState newLifecycleState = LifecycleExecutionState
-            .fromIndexMetadata(newClusterState.metaData().index(index));
-        LifecycleExecutionState oldLifecycleState = LifecycleExecutionState
-            .fromIndexMetadata(oldClusterState.metaData().index(index));
-        assertNotSame(oldLifecycleState, newLifecycleState);
-        assertEquals(currentStep.getPhase(), newLifecycleState.getPhase());
-        assertEquals(currentStep.getAction(), newLifecycleState.getAction());
-        assertEquals(currentStep.getName(), newLifecycleState.getStep());
-        assertEquals(expectedstepInfoValue, newLifecycleState.getStepInfo());
-        assertEquals(oldLifecycleState.getPhaseTime(), newLifecycleState.getPhaseTime());
-        assertEquals(oldLifecycleState.getActionTime(), newLifecycleState.getActionTime());
-        assertEquals(newLifecycleState.getStepTime(), newLifecycleState.getStepTime());
     }
 
     static class MockAsyncActionStep extends AsyncActionStep {
@@ -1830,7 +1030,7 @@ public class IndexLifecycleRunnerTests extends ESTestCase {
 
     }
 
-    private static final class RetryableMockStep extends MockStep {
+    static final class RetryableMockStep extends MockStep {
 
         RetryableMockStep(StepKey stepKey, StepKey nextStepKey) {
             super(stepKey, nextStepKey);
@@ -1840,5 +1040,50 @@ public class IndexLifecycleRunnerTests extends ESTestCase {
         public boolean isRetryable() {
             return true;
         }
+    }
+
+    /** A real policy steps registry where getStep can be overridden so that JSON doesn't have to be parsed */
+    public static class MockPolicyStepsRegistry extends PolicyStepsRegistry {
+        private BiFunction<IndexMetaData, StepKey, Step> fn = null;
+        private static Logger logger = LogManager.getLogger(MockPolicyStepsRegistry.class);
+
+        MockPolicyStepsRegistry(SortedMap<String, LifecyclePolicyMetadata> lifecyclePolicyMap, Map<String, Step> firstStepMap,
+                                Map<String, Map<StepKey, Step>> stepMap, NamedXContentRegistry xContentRegistry, Client client) {
+            super(lifecyclePolicyMap, firstStepMap, stepMap, xContentRegistry, client);
+        }
+
+        public void setResolver(BiFunction<IndexMetaData, StepKey, Step> fn) {
+            this.fn = fn;
+        }
+
+        @Override
+        public Step getStep(IndexMetaData indexMetaData, StepKey stepKey) {
+            if (fn == null) {
+                logger.info("--> retrieving step {}", stepKey);
+                return super.getStep(indexMetaData, stepKey);
+            } else {
+                logger.info("--> returning mock step");
+                return fn.apply(indexMetaData, stepKey);
+            }
+        }
+    }
+
+    public static MockPolicyStepsRegistry createOneStepPolicyStepRegistry(String policyName, Step step) {
+        return createMultiStepPolicyStepRegistry(policyName, Collections.singletonList(step));
+    }
+
+    public static MockPolicyStepsRegistry createMultiStepPolicyStepRegistry(String policyName, List<Step> steps) {
+        LifecyclePolicy policy = new LifecyclePolicy(policyName, new HashMap<>());
+        SortedMap<String, LifecyclePolicyMetadata> lifecyclePolicyMap = new TreeMap<>();
+        lifecyclePolicyMap.put(policyName, new LifecyclePolicyMetadata(policy, new HashMap<>(), 1, 1));
+        Map<String, Step> firstStepMap = new HashMap<>();
+        firstStepMap.put(policyName, steps.get(0));
+        Map<String, Map<StepKey, Step>> stepMap = new HashMap<>();
+        Map<StepKey, Step> policySteps = new HashMap<>();
+        steps.forEach(step -> policySteps.put(step.getKey(), step));
+        stepMap.put(policyName, policySteps);
+        Client client = mock(Client.class);
+        when(client.settings()).thenReturn(Settings.EMPTY);
+        return new MockPolicyStepsRegistry(lifecyclePolicyMap, firstStepMap, stepMap, REGISTRY, client);
     }
 }
